refactor: Moved automapper classes to make space for actual services
